fix(core): update binding activation to invoke missing handlers
